backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0  # PostgreSQL async driver
alembic==1.14.0  # Database migrations
psycopg2-binary==2.9.10  # PostgreSQL adapter

# Authentication & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.7.1
cryptography==44.0.0

# HTTP & API
httpx==0.28.1
requests-oauthlib==2.0.0

# OAuth Integrations
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.155.0
msal==1.31.1
notion-client==2.2.1

# Translation
googletrans==4.0.0rc1
deepl==1.19.1

# Monitoring & Logging
prometheus-fastapi-instrumentator==7.0.0
sentry-sdk==2.19.2

# Rate Limiting
slowapi==0.1.9

# Caching
redis==5.2.1
aioredis==2.0.1

# Background Tasks
apscheduler==3.10.4

# Real-time Communication
python-socketio==5.11.4
python-engineio==4.10.1

# Validation
pydantic==2.10.4
email-validator==2.2.0

# Firebase
firebase-admin==6.6.0
